@@ -1,10 +1,12 @@
 import 'package:redux/redux.dart';
+import 'package:tgg/containers/aws_uploader/aws_upload_action.dart';
 import 'package:tgg/containers/waypoints/submissions/submission_types.dart';
 import 'package:tgg/containers/waypoints/submissions/validate/submissions_validator.dart';
 import 'package:tgg/containers/waypoints/waypoint/waypoint_actions.dart';
 import 'package:tgg/containers/waypoints/waypoint/waypoint_state.dart';
 import 'package:tgg/containers/waypoints/waypoint/waypoint_submission_item.dart';
 import 'package:tgg/containers/waypoints/waypoints_actions.dart';
+import 'package:tgg/data/providers.dart';
 import 'package:tgg/redux_model/app_state.dart';
 
 List<Middleware<AppState>> createWaypointMiddleware() {
@@ -12,6 +14,8 @@ List<Middleware<AppState>> createWaypointMiddleware() {
     new TypedMiddleware<AppState, WaypointsCompletedLoadingAction>(
         _initWaypoint()),
     new TypedMiddleware<AppState, WaypointSubmit>(_submit()),
+    new TypedMiddleware<AppState, WaypointStarted>(_trackStarted()),
+    new TypedMiddleware<AppState, WaypointUpdateAnswer>(_updateAnswer()),
   ];
 }
 
@@ -29,7 +33,7 @@ Middleware<AppState> _initWaypoint() {
 }
 
 Middleware<AppState> _submit() {
-  return (Store store, action, NextDispatcher next) async {
+  return (Store<AppState> store, action, NextDispatcher next) async {
     if (action is WaypointSubmit) {
       final WaypointState waypoint = store.state.waypointState;
 
@@ -41,7 +45,49 @@ Middleware<AppState> _submit() {
         store.dispatch(WaypointShowError(error, item.submission));
       });
 
-      if (errors == 0) {}
+      if (errors == 0) {
+        await waypointsRepo.submitAnswer(
+            waypoint.waypoint.id,
+            waypoint.items.map((i) => {
+                  i.submission.type: i.answer,
+                }),
+            waypoint.items.map((i) => i.media).where((i) => i != null));
+
+        store.dispatch(WaypointsStartLoadAction());
+      }
+    }
+    next(action);
+  };
+}
+
+Middleware<AppState> _trackStarted() {
+  return (Store store, action, NextDispatcher next) async {
+    if (action is WaypointStarted) {
+      await waypointsRepo.trackStart(action.waypointId);
+    }
+    next(action);
+  };
+}
+
+Middleware<AppState> _updateAnswer() {
+  return (Store store, action, NextDispatcher next) async {
+    if (action is WaypointUpdateAnswer) {
+      String answer = action.answer;
+      String media;
+      if (SubmissionTypeHelper.isMediaFromString(action.submission.type)) {
+        final AppState state = store.state;
+        String key = mediaRepo.getKey(
+          action.answer,
+          state.playthrough,
+          state.loginResponse.team,
+          state.waypointState.waypoint,
+        );
+        store.dispatch(AddFileToUploadAction(action.answer, key));
+        answer = key;
+        media = key;
+      }
+
+      store.dispatch(WaypointSaveAnswer(answer, action.submission, media));
     }
     next(action);
   };
