@@ -1,21 +1,31 @@
-import 'package:geolocator/geolocator.dart';
+import 'package:meta/meta.dart';
+import 'package:rxdart/rxdart.dart';
 import 'package:tgg/data/login_repo.dart';
+import 'package:tgg/data/providers/location_provider.dart';
 import 'package:tgg/data/providers/providers.dart';
 
 class LocationRepo {
+  final ApiProvider apiProvider;
   final LoginRepo loginRepo;
+  final LocationProvider locationProvider;
 
-  const LocationRepo(this.loginRepo);
+  const LocationRepo({
+    @required this.apiProvider,
+    @required this.loginRepo,
+    @required this.locationProvider,
+  });
 
-  Future sendLocation(Position position) async {
-    final login = await loginRepo.getLoginResponse();
-    final teamId = login.team.id;
-    return apiProvider.location(
-        teamId: teamId,
-        longitude: position.longitude,
-        latitude: position.latitude,
-        accuracy: position.accuracy.toInt());
+  void sendLocation() async {
+    Observable.zip2(
+        Observable.fromFuture(locationProvider.getLocation()),
+        Observable.fromFuture(loginRepo.getLoginResponse())
+            .map((login) => login.team.id),
+        (position, teamId) => apiProvider.location(
+            teamId: teamId,
+            longitude: position.longitude,
+            latitude: position.latitude,
+            accuracy: position.accuracy.toInt())).listen((r) {
+      r.then((_) {});
+    });
   }
 }
-
-final locationRepo = LocationRepo(null);
