@@ -15,12 +15,13 @@ Future<String> apiRequest(
   if (token != null) request.headers.set('Authorization', 'Token $token');
   sendParams(request, params, method);
   HttpClientResponse response = await request.close();
+  print("request: $url \n ${json.encode(params)} \n ${request.headers}");
   try {
     final reply = await response.transform(utf8.decoder).join();
+    print("response ${response.statusCode} : $reply");
     if (response.statusCode < 400) {
       return reply;
     } else {
-      print("network error: $reply");
       if (isJson(reply)) {
         final String errors =
             Map<String, dynamic>.from(json.decode(reply)).values.expand((v) {
