@@ -2,11 +2,9 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:firebase_analytics/observer.dart';
 import 'package:firebase_crashlytics/firebase_crashlytics.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:flutter_redux_navigation/flutter_redux_navigation.dart';
 import 'package:redux/redux.dart';
@@ -20,11 +18,9 @@ import 'package:tgg/middleware/post_location_middleware.dart';
 import 'package:tgg/reducers/app_reducer.dart';
 import 'package:tgg/redux_model/app_state.dart';
 
-import 'bloc/auth/authentication.dart';
-import 'bloc/theme/theme.dart';
+import 'containers/aws_uploader/aws_upload_middleware.dart';
 import 'containers/camera/camera_container.dart';
 import 'containers/waypoints/waypoint/waypoint_middleware.dart';
-import 'data/playthrought_repository.dart';
 import 'data/providers/remote_config.dart';
 import 'data/simple_bloc_delegate.dart';
 import 'ui/auth/login_page.dart';
@@ -55,6 +51,7 @@ class ReduxApp extends StatelessWidget {
         ..addAll(createCameraMiddleware())
         ..addAll(createWaypointsMiddleware())
         ..addAll(createWaypointMiddleware())
+        ..addAll(createUploadMiddleware())
         ..add(LoggingMiddleware.printer())
         ..add(NavigationMiddleware<AppState>()));
 
@@ -92,77 +89,6 @@ MaterialPageRoute _buildRoute(RouteSettings settings, Widget builder) {
   );
 }
 
-class App extends StatelessWidget {
-  final PlaythroughRepo gameRepo;
-  final FirebaseMessaging firebaseMessaging;
-
-  final FirebaseAnalytics analytics;
-
-  App({Key key, this.gameRepo, this.firebaseMessaging, this.analytics})
-      : assert(gameRepo != null),
-        super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    FlutterError.onError = Crashlytics.instance.recordFlutterError;
-
-    firebaseMessaging.configure(
-      onMessage: (Map<String, dynamic> message) async {
-        print("onMessage: $message");
-        //_showItemDialog(message);
-      },
-      onBackgroundMessage: myBackgroundMessageHandler,
-      onLaunch: (Map<String, dynamic> message) async {
-        print("onLaunch: $message");
-        //_navigateToItemDetail(message);
-      },
-      onResume: (Map<String, dynamic> message) async {
-        print("onResume: $message");
-      },
-    );
-
-    return BlocBuilder<ThemeBloc, ThemeChangeState>(builder: (context, state) {
-      return MaterialApp(
-        title: 'Tgg Demo',
-        theme: (state is ThemeChangedState) ? state.theme : ThemeData(),
-        navigatorObservers: [
-          FirebaseAnalyticsObserver(analytics: analytics),
-        ],
-        home: BlocBuilder<AuthenticationBloc, AuthenticationState>(
-            // ignore: missing_return
-            builder: (context, authState) {
-          if (authState is AuthenticationUninitialized)
-            return SplashPage();
-          else if (authState is AuthenticationAuthenticated)
-            return HomePage();
-          else if (authState is AuthenticationUnauthenticated)
-            return LoginPage();
-        }),
-        onGenerateRoute: (RouteSettings settings) {
-          switch (settings.name) {
-            case CameraContainer.routeName:
-              return new MyCustomRoute(
-                builder: (_) => CameraContainer(),
-                settings: settings,
-              );
-            case PreviewPage.routeName:
-              return new MyCustomRoute(
-                builder: (_) => PreviewPage(),
-                settings: settings,
-              );
-            case PreviewPage.routeName:
-              return new MyCustomRoute(
-                builder: (_) => PreviewPage(),
-                settings: settings,
-              );
-          }
-          throw AssertionError("unsupported ${settings.name}");
-        },
-      );
-    });
-  }
-}
-
 Future<dynamic> myBackgroundMessageHandler(Map<String, dynamic> message) {
   print(message);
   if (message.containsKey('data')) {
@@ -176,20 +102,4 @@ Future<dynamic> myBackgroundMessageHandler(Map<String, dynamic> message) {
   }
 
   // Or do other work.
-}
-
-class MyCustomRoute<T> extends MaterialPageRoute<T> {
-  MyCustomRoute({WidgetBuilder builder, RouteSettings settings})
-      : super(builder: builder, settings: settings);
-
-  @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child) {
-    return super
-        .buildTransitions(context, animation, secondaryAnimation, child);
-    //if (settings.isInitialRoute) return child;
-    // Fades between routes. (If you don't want any animation,
-    // just return child.)
-    // return child;//new FadeTransition(opacity: animation, child: child);
-  }
 }
