@@ -1,3 +1,4 @@
+import 'package:tgg/containers/aws_uploader/aws_upload_reducer.dart';
 import 'package:tgg/containers/camera/camera_reducer.dart';
 import 'package:tgg/containers/waypoints/waypoint/waypoint_reducer.dart';
 import 'package:tgg/containers/waypoints/waypoints_reducer.dart';
@@ -7,7 +8,6 @@ import 'package:tgg/reducers/login_reducer.dart';
 import 'package:tgg/reducers/playthrough_reducer.dart';
 import 'package:tgg/reducers/post_location_reducer.dart';
 import 'package:tgg/reducers/theme_data_reducer.dart';
-import 'package:tgg/reducers/upload_reducer.dart';
 import 'package:tgg/redux_model/app_state.dart';
 
 AppState appReducer(AppState state, action) {
@@ -19,7 +19,7 @@ AppState appReducer(AppState state, action) {
     postLocationTimer: postLocationReducer(state.postLocationTimer, action),
     loginState: loginReducer(state.loginState, action),
     homePageState: homePageReducer(state.homePageState, action),
-    uploadFilesState: uploadFilesReducer(state.uploadFilesState, action),
+    uploadFilesState: awsUploadFilesReducer(state.uploadFilesState, action),
     cameraState: cameraReducer(state.cameraState, action),
     waypointsState: waypointsReducer(state.waypointsState, action),
     waypointState: waypointReducer(state.waypointState, action),
