@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
 import 'package:tgg/actions/home_actions.dart';
+import 'package:tgg/containers/aws_uploader/aws_upload_container.dart';
 import 'package:tgg/containers/home/post_location_container.dart';
-import 'package:tgg/containers/home/upload_container.dart';
 import 'package:tgg/models/modes.dart';
 import 'package:tgg/redux_model/app_state.dart';
 import 'package:tgg/ui/keys.dart';
