@@ -1,6 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:geolocator/geolocator.dart';
 import 'package:mockito/mockito.dart';
 import 'package:tgg/data/providers/api_provider.dart';
+import 'package:tgg/data/providers/location_provider.dart';
 import 'package:tgg/data/providers/prefs_provider.dart';
 import 'package:tgg/data/providers/providers.dart';
 
@@ -9,6 +11,8 @@ class MockedApiProvider extends Mock implements ApiProvider {}
 class MockedStaticApiProvider extends Mock implements StaticApiProvider {}
 
 class MockedPrefsProvider extends Mock implements PrefsProvider {}
+
+class MockedLocationProvider extends Mock implements LocationProvider {}
 
 const goodPin = "good";
 const badPin = "bad";
@@ -24,11 +28,15 @@ const mockedFlavor =
     """{"primary.color":"#f47421","success.color":"#46a546","danger.color":"#9d261d","background.color":"#efeeec","foreground.color":"#333333","backgroundAlt.color":"#ffffff","foregroundAlt.color":"#333333","foregroundInverse.color":"#ffffff","headerBackground.color":"#ffffff","navBarBackground.color":"#ffffff","font.color":"#000000","facebook.color":"#3b5998","twitter.color":"#4099FF","nav.header.image":"","nav.header.align":"center","nav.footer.image":"","nav.footer.align":"center","global.loading":"Loading...","global.unavailable":"Unavailable","global.notloggedin":"You aren't logged in.","global.refresh":"Refresh","global.with":"with","global.refreshing":"Refreshing...","global.pts":"pts","global.points":"points","global.nomissions":"No missions available","global.Cancel":"Cancel","intro.loading":"Routing...","intro.play":"Go!","intro.welcome":"Welcome to the Go Game! Get psyched!","intro.closed":"The game will begin soon.","intro.title":"The Go Game","intro.realtime.testing":"Testing realtime notifications...","intro.realtime.failure":"Realtime is offline!","intro.realtime.restart":"Restart to resolve","intro.realtime.retest":"Retest","login.notstarted":"The game hasn't started yet. Please check again in a little while!","login.login":"Log in","login.logging_in":"Logging in...","login.invalid_email":"Enter a valid email address","login.invalid_password":"Incorrect password","login.error":"Error logging in.","signup.sign_up":"Sign up!","signup.signing_up":"Signing up...","signup.email_sent":"Success!","signup.check_email":"We just sent an email to that address.\n Check your inbox to get started!","signup.email_registered":"This email address is already signed up!\n Check your email for your login information.","signup.resend":"Resend","nav.map":"Map","nav.play":"Play","nav.modes.camera":"Bonus","nav.modes.questions":"Trivia","nav.modes.main":"Play","nav.modes.head_to_head":"H2H","nav.modes.anytime":"Anytime","nav.admin":"Admin","nav.inbox":"Inbox","nav.team":"Team","nav.vote":"Vote","nav.home":"Home","nav.login":"Login","nav.select_votables":"Select","nav.vote_admin":"Voting","nav.info":"Info","nav.leaderboard":"Leaders","nav.gallery":"Gallery","nav.modes.intro":"Intro","nav.modes.sneak":"Sneak","nav.modes.keynote":"Keynote","nav.modes.creative":"Cam","nav.modes.vendor":"Vendor","sync.failure":"Upload failure.","sync.success":"All done!","home.paused.message":"The game is paused! You will receive more information shortly. Until then, take a break and rehearse your victory speech","home.paused.title":"Game paused","home.ended.defaultmessage":"The game is over! Get yourself to your ending venue and we'll start the judging shortly.","home.ended.title":"Game Over","home.ended.title_early":"You've done it!","home.inprogress.time_remaining":"remaining in game.","home.inprogress.title":"You're on a roll!","home.login.camera.message":"You can log in your personal phones as game cameras! Go to {{url}} and login with the code {{code}}.","home.vote":"Start voting","home.vote_select":"Please select missions for voting.","home.vote_select_button":"Select missions","home.logout":"Exit game","home.logout.confirm_message":"Are you sure you want to exit?","home.logout.confirm_title":"Please confirm","home.inprogress.gated_title":"Take a Short Break","home.inprogress.gated_message":"OK, {{teamname}}! You are doing great! While we take a moment to spy on the other teams, feel free to find another group and do a Head-To-Head mission (for massive points!) or take a break to grab some refreshments of your choice (for no points, but other benefits!).","home.reentered.message":"You're not done yet! Your Game Producer has enabled some more missions, hit \"Continue game\" to keep playing","modes.camera.title":"Bonus Cam","modes.anytime.title":"Anytime Missions","modes.anytime.home.continue":"Play the game","modes.anytime.home.routing":"Routing...","modes.anytime.home.error":"Routing error","modes.questions.home.continue":"Next Question","modes.questions.home.routing":"Finding your next question!","modes.questions.home.error":"Please try again!","modes.questions.title":"Next Question","modes.main.home.continue":"Continue game","modes.main.home.routing":"Finding your next mission!","modes.main.home.error":"Please try again!","modes.main.title":"Continue game","modes.head_to_head.home.respond":"Respond","modes.head_to_head.home.your_pin":"Your code is: {{pin}}","modes.head_to_head.home.play":"Continue mission","modes.head_to_head.home.pin_challenge_with_remaining":"Challenge ({{numRemaining}} left)","modes.head_to_head.home.received":"{{teamName}} has challenged you!","modes.head_to_head.home.pin_placeholder":"Enter code","modes.head_to_head.home.uninterruptible":"You cannot send challenges while you are in the middle of an actor mission. Finish it and then try again.","modes.head_to_head.home.sent":"Waiting for response from {{teamName}}.","modes.head_to_head.home.pin_prompt":"Head to Heads are worth up to 100 points! Enter another team's code to challenge them. Your code is {{pin}}.","modes.head_to_head.home.inprogress":"You have a Head to Head in progress","modes.head_to_head.home.pin_challenge":"Send challenge","modes.head_to_head.home.resend":"Resend","modes.head_to_head.inprogress.sending":"Sending...","modes.head_to_head.inprogress.error":"Sorry there was an error. Try again?","modes.head_to_head.inprogress.declining":"Declining...","modes.head_to_head.inprogress.accepting":"Accepting...","modes.head_to_head.confirm.other_uninterruptible":"{{teamName}} is in the middle of a actor mission. Try again after they've finished.","modes.head_to_head.confirm.received":"You've been challenged by {{teamName}}!","modes.head_to_head.confirm.team_not_found":"Team not found.","modes.head_to_head.confirm.did_decline":"Declined.","modes.head_to_head.confirm.did_accept":"You accepted this challenge!","modes.head_to_head.confirm.other_team_default":"the other team","modes.head_to_head.confirm.self_reached_max_total":"Sorry, you've reached your total maximum challenges.","modes.head_to_head.confirm.exhausted":"Sorry, you have completed all the missions we have with this team! Try challenging someone else.","modes.head_to_head.confirm.was_declined":"{{teamName}} declined your challenge. They might have been busy - feel free to try again soon!","modes.head_to_head.confirm.reached_max_per_team":"Sorry, you've reached your max number of challenges with {{teamName}}.","modes.head_to_head.confirm.expired":"Sorry, this challenge expired. Try again!","modes.head_to_head.confirm.other_reached_max_total":"Sorry, {{teamName}} has reached their total maximum challenges.","modes.head_to_head.confirm.challenge_inactive":"This challenge isn't active right now. Wait a bit, then try again.","modes.head_to_head.confirm.was_accepted":"{{teamName}} accepted your challenge! Your current mission has been paused; you'll resume it when this is done.","modes.head_to_head.confirm.received_footer":"Only accept if you are physically near {{teamName}}.","modes.head_to_head.confirm.other_busy":"{{teamName}} is busy. Wait a bit and try again.","modes.head_to_head.confirm.self_uninterruptible":"You can't make challenges while you're in the middle of a actor mission. Finish your mission and try again.","modes.head_to_head.confirm.self_busy":"Sorry, you actually have more important missions to complete right now. Hit the Play tab to keep going, and try a H2H later!","modes.head_to_head.confirm.not_self":"You can't challenge yourself.","modes.head_to_head.confirm.sent":"Sent!","modes.head_to_head.actions.decline":"Decline","modes.head_to_head.actions.accept":"Accept","modes.head_to_head.title":"Head to Head","persistent.media_warning":"You currently have a photo or video mission in progress. You should take mission photos here:","mission.incorrect.tryagain.0.text":"McKayla is not impressed. {{given}} is incorrect. Try again!","mission.incorrect.tryagain.0.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f0mckayla.jpg","mission.incorrect.tryagain.1.text":"{{given}} is incorrect. Try again.","mission.incorrect.tryagain.1.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f1grumpycat.jpg","mission.incorrect.tryagain.2.text":"Nah. {{given}} is incorrect. Try again.","mission.incorrect.tryagain.2.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f2obama.gif","mission.incorrect.tryagain.3.text":"{{given}} is incorrect. Try again.","mission.incorrect.tryagain.3.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f3embarassedpolar.jpg","mission.incorrect.tryagain.4.text":"Nuh-uh. {{given}} is incorrect. Try again.","mission.incorrect.tryagain.4.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f4funnyfall.jpg","mission.incorrect.tryagain.5.text":"Nope. {{given}} is incorrect. Try again.","mission.incorrect.tryagain.5.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f5sadkeanu.jpg","mission.incorrect.tryagain.6.text":"No dice. {{given}} is incorrect. Try again.","mission.incorrect.tryagain.6.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f6nickyoung.gif","mission.incorrect.tryagain.7.text":"{{given}} is incorrect. Try again.","mission.incorrect.tryagain.7.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f7chubbybubbles.jpg","mission.incorrect.tryagain.8.text":"{{given}} is incorrect. Try again.","mission.incorrect.tryagain.8.image":"https://gogame-breadcrumb-content.s3.amazonaws.com/default/fail/sad-owl-fail.jpg","mission.incorrect.final.0.text":"McKayla is not impressed. The right answer is {{correct}}.","mission.incorrect.final.0.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f0mckayla.jpg","mission.incorrect.final.1.text":"The right answer is {{correct}}.","mission.incorrect.final.1.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f1grumpycat.jpg","mission.incorrect.final.2.text":"Nah. The right answer is {{correct}}.","mission.incorrect.final.2.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f2obama.gif","mission.incorrect.final.3.text":"Not so much. The right answer is {{correct}}.","mission.incorrect.final.3.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f3embarassedpolar.jpg","mission.incorrect.final.4.text":"Nuh-uh. The right answer is {{correct}}.","mission.incorrect.final.4.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f4funnyfall.jpg","mission.incorrect.final.5.text":"Nope. The right answer is {{correct}}.","mission.incorrect.final.5.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f5sadkeanu.jpg","mission.incorrect.final.6.text":"No dice. The right answer is {{correct}}.","mission.incorrect.final.6.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f6nickyoung.gif","mission.incorrect.final.7.text":"The right answer is {{correct}}.","mission.incorrect.final.7.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/fail/f7chubbybubbles.jpg","mission.incorrect.final.8.text":"The right answer is {{correct}}.","mission.incorrect.final.8.image":"https://gogame-breadcrumb-content.s3.amazonaws.com/default/fail/sad-owl-fail.jpg","mission.correct.0.text":"","mission.correct.0.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s0bravocado.jpg","mission.correct.1.text":"","mission.correct.1.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s1shaqcat.gif","mission.correct.2.text":"Smooth moves","mission.correct.2.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s2biglebowski.gif","mission.correct.3.text":"","mission.correct.3.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s3carlton.gif","mission.correct.4.text":"Too cool for school","mission.correct.4.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s4peggyolson.gif","mission.correct.5.text":"Put ‘er there!","mission.correct.5.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s5superman.jpg","mission.correct.6.text":"Daaamn!","mission.correct.6.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s6damn.gif","mission.correct.7.text":"Rock on","mission.correct.7.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s7bettywhite.jpg","mission.correct.8.text":"","mission.correct.8.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/s8homer.gif","mission.skipped":"Your game runner routed you to a different mission -- onwards!","mission.ok.0.text":"","mission.ok.0.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/o0missionaccomplished.gif","mission.ok.1.text":"","mission.ok.1.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/o1sloth.jpg","mission.ok.2.text":"You slay.","mission.ok.2.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/o2beyslay.jpg","mission.ok.3.text":"","mission.ok.3.image":"https://s3.amazonaws.com/gogame-breadcrumb-content/default/win/o3dolphinsmiles.jpg","mission.timeout.expired":"Overtime!","mission.timeout.remaining":"Time left:","mission.captionrequired":"A caption is required","mission.abort":"Nevermind!","mission.skip":"Skip","mission.wait_refresh":"Refresh","mission.footnotes.linked_head_to_head":"Winner gets full points and the loser gets half. If you can't agree, both teams lose points!","mission.upload.native_pending":"Getting media...","mission.upload.mobile":"Tap to take {{mediaType}}","mission.upload.mobile.optional":"Optional {{mediaType}}","mission.upload.desktop":"Drag {{mediaType}} here","mission.upload.desktop.optional":"Optional {{mediaType}}","mission.upload.invalidtype":"Please try another file type.","mission.upload.failed":"Upload failed.","mission.upload.tip.video.android":"WARNING: Reduce your video resolution to 640x480, or your videos will be too large to upload!","mission.upload.tip.photo":"PROTIP: Turn your phone horizontal for the best photos.","mission.upload.tip.landscape":"Remember to take photos in LANDSCAPE so they will look awesome on the big screen!","mission.upload.toolarge":"Upload is too large! Try making a shorter video and trying again.","mission.upload.ready":"Ready to upload!","mission.upload.retry":"Discard and Retry","mission.upload.native_add_error":"Failed to add media. Make sure the Go Game app has permission to access your photos.","mission.upload.native_media_cancelled":"Taking media cancelled.","mission.submit":"Go","mission.attempts_remaining":"Tries left: {{numAttemptsRemaining}}","mission.blank.0":"Please provide an answer.","mission.code.ok":"Excellent!","mission.code.invalid.tryagain":"That code was invalid - try again!","mission.code.invalid.final":"That code was invalid - time to move on!","mission.code.blank":"Please enter a code","mission.submitting":"Submitting...","mission.error":"There was an error making this submission. Try again?","mission.wait_refreshing":"Refreshing...","mission.linked.disagreed":"You didn't agree! Both teams lose points!","mission.linked.won":"Nice work team, you won!","mission.linked.ok":"Got it!","mission.linked.lost":"You lost - better luck next time, but you still got some points!","mission.linked.wait":"Waiting for rest of teams...","mission.hints.get":"Get hint for {{hintPointCost}} pts ({{numRemaining}} left)","modes.intro.title":"Intro Missions","modes.sneak.title":"Location Missions","modes.keynote.title":"Keynote Missions","modes.creative.title":"Creative Missions","modes.vendor.title":"Vendor Missions","message.send":"Send","message.cancel":"Cancel","message.yes":"Yes","message.ok":"Ok!","message.no":"No","message.verify.info":"After reading this message, enter '{{verify}}' below to continue.","message.verify.confirm":"Got it","social.logout":"Log out","social.share":"Share +{{pointsForSharing}} pts","social.message.nopoints":"Share","social.message.default":"We're playing #TheGoGame! www.thegogame.com","social.skip":"Don't Share","social.select":"Share for Points","social.instructions":"Select the social network you would like to use below. You can always change your mind later.","social.saving":"Saving Social Selection...","social.facebook.name":"Facebook","social.facebook.icon":"icon-facebook-sign","social.facebook.waiting":"Waiting for response from Facebook...","social.linkedin.name":"LinkedIn","social.linkedin.icon":"icon-linkedin-sign","social.linkedin.waiting":"Waiting for response from LinkedIn...","social.twitter.name":"Twitter","social.twitter.icon":"icon-twitter-sign","social.twitter.waiting":"Waiting for response from Twitter...","persistent.queue_error":"There was an error loading your mission. Try again?","persistent.nomissions":"No missions available.","leaderboard.not_shown":"teams not shown","voting.inprogress":"Sending vote","voting.no_waypoint":"Voting hasn't started yet.","voting.own_waypoint":"You can't vote on your own mission.","voting.updated":"Your vote was updated.","voting.received":"Your vote was received!","voting.score4":"4: A-OK","voting.score5":"5: Pretty great","voting.score6":"6: Perfect 10","voting.score1":"1: Piss-poor","voting.score2":"2: Meh","voting.score3":"3: All right","voting.select.notification":"Before continuing with the game, please choose which creative missions you'd like to feature in the voting phase of the game.","voting.select.none_to_select":"None to select","voting.select.ok":"Great, you're all set!","voting.select.notready":"Currently {{numChosen}}","voting.select.confirm":"Confirm","voting.select.choose":"Choose {{numToChoose}} missions for voting."}""";
 
 const mockedActiveWaypoints =
-    """[{"id":"5e0878bcd71f1000192aae2e","step":{"id":"5e072f3fd71f10001a444b38","title":"Name your team","instructions":"give a name for your team!","position":0,"mission_id":"5e072f19d71f10001a444b37","behavior":{"id":"profile_team_name","title":"Profile: Team Name","description":"Team's submission for this mission is assigned as their team name.","fields":[{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"teamname_placeholder","type":"text","label":"Placeholder for TeamName field","default":"Team name","required":false,"placeholder":"Team Name"},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":{"type":"text","default_placeholder":"Team name","placeholder":"content.teamname_placeholder"}},"content":{"success_message":"Brilliant!"},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2019-12-29T09:58:20Z","queue_priority":0,"ungated_at":"2019-12-29T09:58:20Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]},{"id":"5e0878bcd71f1000192aae2f","step":{"id":"5e072f87d71f10001a444b3a","title":"Selfie time!","instructions":"Make a picture of your team","position":0,"mission_id":"5e072f87d71f10001a444b39","behavior":{"id":"profile_photo","title":"Profile: Team Photo","description":"Teams's photo submission is set to their team photo.","fields":[{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"select_gallery_media_enabled","type":"bool","default":false,"label":"Allow submission from existing media","required":false},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":"photo"},"content":{"success_message":"You rock!"},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2019-12-29T09:58:20Z","queue_priority":0,"ungated_at":"2019-12-29T09:58:20Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]}]""";
+    """[{"id":"5e274618d71f10001e0e4584","step":{"id":"5e072f3fd71f10001a444b38","title":"Name your team","instructions":"give a **name** for *your* team!","position":0,"mission_id":"5e072f19d71f10001a444b37","behavior":{"id":"profile_team_name","title":"Profile: Team Name","description":"Team's submission for this mission is assigned as their team name.","fields":[{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"teamname_placeholder","type":"text","label":"Placeholder for TeamName field","default":"Team name","required":false,"placeholder":"Team Name"},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":{"type":"text","default_placeholder":"Team name","placeholder":"content.teamname_placeholder"}},"content":{"success_message":"Brilliant!"},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2020-01-21T18:42:32Z","queue_priority":0,"ungated_at":"2020-01-21T18:42:32Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]},{"id":"5e274618d71f10001e0e4585","step":{"id":"5e072f87d71f10001a444b3a","title":"Selfie time!","instructions":"Make a picture of your team","position":0,"mission_id":"5e072f87d71f10001a444b39","behavior":{"id":"profile_photo","title":"Profile: Team Photo","description":"Teams's photo submission is set to their team photo.","fields":[{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"select_gallery_media_enabled","type":"bool","default":false,"label":"Allow submission from existing media","required":false},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":"photo"},"content":{},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2020-01-21T18:42:32Z","queue_priority":0,"ungated_at":"2020-01-21T18:42:32Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]},{"id":"5e274618d71f10001e0e4586","step":{"id":"5e25feedd71f10001e0e4559","title":"phone","instructions":"enter your phone number","position":0,"mission_id":"5e25feb1d71f10001e0e4558","behavior":{"id":"profile_phone","title":"Profile: Phone Number","description":"Team's submission is logged as their phone number.","fields":[{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":[{"type":"number","placeholder":"Phone number"}]},"content":{},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2020-01-21T18:42:32Z","queue_priority":0,"ungated_at":"2020-01-21T18:42:32Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]},{"id":"5e274619d71f10001e0e4587","step":{"id":"5e25ff4bd71f10001e0e455b","title":"select correct","instructions":"Select correct answer","position":0,"mission_id":"5e25ff15d71f10001e0e455a","behavior":{"id":"multiple_choice","title":"Multiple Choice","description":"Choose one from a multiple of choices, only some of which are correct.","fields":[{"name":"choices","type":"list:text+check","label":"Options","required":true},{"name":"hints","type":"list:text","label":"Hints","required":false},{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"failure_message","type":"text","label":"Fail message","required":false},{"name":"timelimit","type":"int","label":"Time limit","default":null,"required":false,"placeholder":"Time limit (in minutes)"},{"name":"num_attempts","type":"int","default":3,"label":"Number of attempts","required":false,"placeholder":"Number of attempts allowed"},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":{"type":"choice","choices":"content.choices"}},"content":{"choices":[{"text":"first","value":false},{"text":"second","value":false},{"text":"correct","value":true},{"text":"last","value":false}]},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2020-01-21T18:42:32Z","queue_priority":0,"ungated_at":"2020-01-21T18:42:32Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]},{"id":"5e274619d71f10001e0e4588","step":{"id":"5e25ffcbd71f10001e0e455d","title":"many letters","instructions":"you can write any long text here","position":0,"mission_id":"5e25ff76d71f10001e0e455c","behavior":{"id":"worksheet","title":"Worksheet","description":"Team's submission for this mission is assigned as their worksheet.","fields":[{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":"text"},"content":{},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2020-01-21T18:42:33Z","queue_priority":0,"ungated_at":"2020-01-21T18:42:33Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]},{"id":"5e274619d71f10001e0e4589","step":{"id":"5e26006dd71f10001e0e4560","title":"Movie and text","instructions":"record a movie and name it","position":0,"mission_id":"5e26004ed71f10001e0e455f","behavior":{"id":"movie_and_text","title":"Movie and Text","description":"Submit a movie and a caption.","fields":[{"name":"success_message","type":"text","label":"Success message","required":false},{"name":"timelimit","type":"int","label":"Time limit","default":null,"required":false,"placeholder":"Time limit (in minutes)"},{"name":"video_max_length","type":"int","default":30,"label":"Max video length","placeholder":"(seconds)","required":false},{"name":"video_quality","type":"select","options":[{"value":"med","label":"Medium (480p)"},{"value":"high","label":"High (720p)"},{"value":"very high","label":"Very High (1080p)"}],"label":"Video quality","default":"med","required":false},{"name":"video_pause_record","type":"bool","default":false,"label":"Pause record","required":false},{"name":"video_disable_rotation","type":"bool","default":false,"label":"Video Disable Rotation","required":false},{"name":"select_gallery_media_enabled","type":"bool","default":false,"label":"Allow submission from existing media","required":false},{"name":"presentation_text","type":"text:long","label":"Presentation text","required":false,"placeholder":"Presentation Text"},{"name":"presentation_image","type":"upload","label":"Presentation image","required":false,"placeholder":"Image key","object_key":"media_key","s3":{"bucket":"gogame-breadcrumb-display-media","folder":"GameShow"}},{"name":"presentation_title","type":"text","label":"Presentation title","required":false,"placeholder":"Presentation Title (overrides mission step title)"},{"name":"display_answer","type":"text","label":"GGP Display answer","required":false,"placeholder":"Correct display answer. Will display on screen view"},{"name":"question_type","type":"text","label":"GGP Question type","required":false,"placeholder":"Question type"},{"name":"timer","type":"int","label":"GGP Timer","required":false},{"name":"stop_auto_play","type":"bool","default":false,"label":"GGP Stop autoplay","required":false},{"name":"is_one_time_mission","type":"bool","default":false,"label":"Is Mission available only once per playthrough","required":false}],"submission_type":["movie",{"type":"text","placeholder":"Caption"}]},"content":{},"scoring":0},"mode_name":null,"step_num":0,"is_last":true,"is_from_start_or_end_track":true,"queued_at":"2020-01-21T18:42:33Z","queue_priority":0,"ungated_at":"2020-01-21T18:42:33Z","gated_until":null,"social_info":{},"has_shared":false,"gated_by_mission_delay":null,"started_at":null,"started_location":null,"completed_at":null,"completed_location":null,"synced_at":null,"submissions":[],"num_hints_used":null,"result":{},"is_votable":false,"is_voting_selected":false,"is_judgeable":true,"votes":[],"points":0,"points_are_overridden":false,"has_media":false,"in_gallery":false,"media":[]}]""";
+
+const mockCreateMedia =
+    """{"media_bucket":"gogame-breadcrumb-display-media","upload_key":"2017-08-12/immaculate-conception-church-powell-11-07/camera-/art-pose-2017-08-12-11-23-38.png","media_key":"2017-08-12/immaculate-conception-church-powell-11-07/camera-/images-<size>/art-pose-2017-08-12-11-23-38.jpeg","upload_mime_type":"image/png","media_mime_type":"image/jpeg","media_timestamp":1502562059,"media_type":"image","media_filename":"photoCompressed5.png","media_size":4616904,"options":{"is_uploaded":true,"is_uploading":true,"in_gallery":true,"is_ready":true,"media_duration":4,"caption":"Donkey Kong"},"id":"5c743270cffcee7eef899092"}""";
 
 final mockedApiProvider = MockedApiProvider();
 final mockedStaticApiProvider = MockedStaticApiProvider();
 final mockedPrefs = MockedPrefsProvider();
+final mockedLocationProvider = MockedLocationProvider();
 
 void initLoginMock() {
   when(mockedApiProvider.login(goodPin))
@@ -39,10 +47,24 @@ void initLoginMock() {
       .thenAnswer((_) => Future.value(mockedPlaythroughs));
   when(mockedApiProvider.activeWaypoints())
       .thenAnswer((_) => Future.value(mockedActiveWaypoints));
+  when(mockedApiProvider.waypointTriggerAction(
+    waypointId: anyNamed("waypointId"),
+    values: anyNamed("values"),
+  )).thenAnswer((_) => Future.value(mockedActiveWaypoints));
+  when(mockedApiProvider.createMedia(values: anyNamed("values")))
+      .thenAnswer((_) => Future.value(mockCreateMedia));
+  when(mockedApiProvider.putMedia(
+    id: anyNamed("id"),
+    values: anyNamed("values"),
+  )).thenAnswer((_) => Future.value(mockCreateMedia));
 
   //static
   when(mockedStaticApiProvider.fetchBlueprint())
       .thenAnswer((_) => Future.value(mockedBlueprint));
   when(mockedStaticApiProvider.fetchFlavor())
       .thenAnswer((_) => Future.value(mockedFlavor));
+
+  final position = Position.fromMap({'latitude': 50.44, 'longitude': 30.39});
+  when(mockedLocationProvider.getLocation())
+      .thenAnswer((_) => Future.value(position));
 }
